# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "区块链研究助手 API",
        "description": "区块链研究助手的 REST API 文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:10000",
            "description": "开发服务器"
        }
    ],
    "paths": {
        "/api/health": {
            "get": {
                "summary": "健康检查",
                "description": "检查 API 服务是否正常运行",
                "responses": {
                    "200": {
                        "description": "服务正常",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "healthy"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/upload": {
            "post": {
                "summary": "上传文件",
                "description": "上传 PDF 文件进行分析",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "文件上传成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "File uploaded successfully"
                                        },
                                        "filename": {
                                            "type": "string",
                                            "example": "example.pdf"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No file selected"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/summarize": {
            "post": {
                "summary": "生成文本摘要",
                "description": "对输入文本生成摘要",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "区块链是一个去中心化的分布式账本技术..."
                                    },
                                    "max_length": {
                                        "type": "integer",
                                        "example": 130
                                    },
                                    "min_length": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "required": ["text"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "摘要生成成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "summary": {
                                            "type": "string",
                                            "example": "区块链是一个去中心化的分布式账本技术..."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/search": {
            "post": {
                "summary": "搜索文档",
                "description": "搜索相似文档",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "区块链技术的应用"
                                    },
                                    "k": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "required": ["query"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "搜索成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "distances": {
                                            "type": "array",
                                            "items": {
                                                "type": "number"
                                            }
                                        },
                                        "indices": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
} 